Disabled test until final definition of view
